requirements.txt
python-dotenv==1.0.*
aiogram==2.25.0
requests==2.31.0